refactor user schema
